refactor(:black_circle:): remove circular dependency in request
AFFECTS PACKAGES:
@esri/arcgis-rest-request
