+++ packages/arcgis-rest-request/src/index.ts
@@ -4,10 +4,11 @@
 export * from "./request";
 export * from "./utils/encode-form-data";
 export * from "./utils/encode-query-string";
-export * from "./utils/check-for-errors";
+export * from "./utils/warn";
 export * from "./utils/ArcGISRequestError";
-export * from "./utils/ArcGISAuthError";
+export * from "./utils/retryAuthError";
 export * from "./utils/ErrorTypes";
+export * from "./utils/params";
 export * from "./utils/process-params";
 export * from "./utils/get-portal";
 export * from "./utils/get-portal-url";
